+++ CHANGELOG.md
@@ -74,10 +74,16 @@ Those changes in added, changed or breaking changes, should include usage exampl
         .await
         .unwrap();
   ```
+  
 * Added new method `set_ejector` in `avsregistry/writer` in [#330](https://github.com/Layr-Labs/eigensdk-rs/pull/330).
   ```rust
     let new_ejector_address = address!("70997970C51812dc3A010C7d01b50e0d17dc79C8");
     let tx_hash = avs_writer.set_ejector(new_ejector_address).await.unwrap();
+  ````
+
+* Added new method `is_operator_set_quorum` in `avsregistry/writer` in [#296](https://github.com/Layr-Labs/eigensdk-rs/pull/296).
+  ```rust
+    let operator_set_quourm = avs_reader.is_operator_set_quorum(0).await.unwrap();
   ```
 
 ### Changed